requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
cycler==0.10.0
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
idna==3.2
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
numpy==1.21.2
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
Pillow==8.3.2
PyMySQL==1.0.2
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.24
typing-extensions==3.10.0.2
urllib3==1.26.6
Werkzeug==2.0.1
zipp==3.5.0
